Divide up unit tests
After refactoring, we move the output tests out of the `StreamTransformCSVService` unit tests and into `ConvertToCSVService` tests
