+++ test/services/streams/stream_transform_csv.service.test.js
@@ -3,8 +3,9 @@
 // Test framework dependencies
 const Lab = require('@hapi/lab')
 const Code = require('@hapi/code')
+const Sinon = require('sinon')
 
-const { describe, it } = exports.lab = Lab.script()
+const { describe, it, beforeEach, afterEach } = exports.lab = Lab.script()
 const { expect } = Code
 
 const stream = require('stream')
@@ -12,10 +13,23 @@ const stream = require('stream')
 // Test helpers
 const { StreamHelper } = require('../../support/helpers')
 
+// Things we need to stub
+const { ConvertToCSVService } = require('../../../app/services')
+
 // Thing under test
 const { StreamTransformCSVService } = require('../../../app/services')
 
 describe('Stream Transform CSV service', () => {
+  let csvStub
+
+  beforeEach(() => {
+    csvStub = Sinon.stub(ConvertToCSVService, 'go')
+  })
+
+  afterEach(() => {
+    Sinon.restore()
+  })
+
   describe('When data is passed to it', () => {
     it('returns a stream', async () => {
       const result = StreamTransformCSVService.go()
@@ -23,32 +37,14 @@ describe('Stream Transform CSV service', () => {
       expect(result).to.be.an.instanceof(stream.Stream)
     })
 
-    it('streams the correct data', async () => {
-      const testData = ['first', '2', 'false']
+    it('passes data to ConvertToCSVService', async () => {
+      const testData = ['test', 'data']
 
       const transformStream = StreamTransformCSVService.go()
-      // We use destructuring to pull the sole element of the array into result
-      const [result] = await StreamHelper.testTransformStream(transformStream, testData)
+      await StreamHelper.testTransformStream(transformStream, testData)
 
-      expect(result).to.equal('"first","2","false"\n')
-    })
-
-    it('correctly formats objects', async () => {
-      const testData = [{ __DecisionID__: '52c1088f-41b7-4f26-a7db-cab9792dc1cb0', WRLSChargingResponse: { chargeValue: 215.28, decisionPoints: { sourceFactor: 9, seasonFactor: 14.4, lossFactor: 14.4, volumeFactor: 1, abatementAdjustment: 215.28, s127Agreement: 215.28, s130Agreement: 215.28, secondPartCharge: false, waterUndertaker: false, eiucFactor: 0, compensationCharge: false, eiucSourceFactor: 0, sucFactor: 215.28 }, messages: [], sucFactor: 14.95, volumeFactor: 1, sourceFactor: 9, seasonFactor: 1.6, lossFactor: 1, abatementAdjustment: 'S126 x 1.0', s127Agreement: null, s130Agreement: null, eiucSourceFactor: 0, eiucFactor: 0 } }]
-
-      const transformStream = StreamTransformCSVService.go()
-      const [result] = await StreamHelper.testTransformStream(transformStream, testData)
-
-      expect(result).to.equal('"{""__DecisionID__"":""52c1088f-41b7-4f26-a7db-cab9792dc1cb0"",""WRLSChargingResponse"":{""chargeValue"":215.28,""decisionPoints"":{""sourceFactor"":9,""seasonFactor"":14.4,""lossFactor"":14.4,""volumeFactor"":1,""abatementAdjustment"":215.28,""s127Agreement"":215.28,""s130Agreement"":215.28,""secondPartCharge"":false,""waterUndertaker"":false,""eiucFactor"":0,""compensationCharge"":false,""eiucSourceFactor"":0,""sucFactor"":215.28},""messages"":[],""sucFactor"":14.95,""volumeFactor"":1,""sourceFactor"":9,""seasonFactor"":1.6,""lossFactor"":1,""abatementAdjustment"":""S126 x 1.0"",""s127Agreement"":null,""s130Agreement"":null,""eiucSourceFactor"":0,""eiucFactor"":0}}"\n')
-    })
-
-    it('correctly handles strings with quotes', async () => {
-      const testData = ['Some "Quotes" Here']
-
-      const transformStream = StreamTransformCSVService.go()
-      const [result] = await StreamHelper.testTransformStream(transformStream, testData)
-
-      expect(result).to.equal('"Some ""Quotes"" Here"\n')
+      expect(csvStub.calledOnce).to.be.true()
+      expect(csvStub.firstCall.firstArg).to.equal(testData)
     })
   })
 })
